se agregan usuarioDTO y topicoDTO, se corrijes sus respectivos controller y test unitarios

+++ src/main/java/com/alura/foro/service/TopicoService.java
@@ -1,16 +1,16 @@
 package com.alura.foro.service;
 
-import com.alura.foro.modelo.Curso;
-import com.alura.foro.modelo.Respuesta;
+import com.alura.foro.dto.responce.TopicoDTO;
 import com.alura.foro.modelo.Topico;
 import com.alura.foro.repository.TopicoRepository;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class TopicoService {
@@ -18,28 +18,47 @@ public class TopicoService {
     @Autowired
     private TopicoRepository repository;
 
+    @Autowired
+    private ModelMapper modelMapper;
 
-    public List<Topico> getTopicos (){
-        return repository.findAll();
+
+    public List<TopicoDTO> getTopicos (){
+        List<Topico> topicos =  repository.findAll();
+
+        return topicos.stream().map(TopicoDTO::new).collect(Collectors.toList());
     }
 
-    public Topico saveTopico (@Valid Topico topico){
-        return repository.save(topico);
+    public TopicoDTO setTopico (@Valid Topico topico){
+
+        Topico saveTopic = repository.save(topico);
+        TopicoDTO topicoDTO = modelMapper.map(saveTopic,TopicoDTO.class);
+        return topicoDTO;
     }
 
-    public void deleteTopico (Long id){
-        repository.deleteById(id);
+    public void deleteTopico (Long id) throws EntityNotFoundException {
+        if(repository.findById(id).isPresent()){
+            repository.deleteById(id);
+        }else {
+            throw new EntityNotFoundException("Tópico no encontrado");
+        }
+
+
     }
 
-    public void updateTopico (Topico topico){
-        repository.save(topico);
+    public void updateTopico (Topico topico) throws EntityNotFoundException{
+        if(repository.findById(topico.getId()).isPresent()) {
+            repository.save(topico);
+        }else {
+            throw new EntityNotFoundException("Tópico a acutlaizar no encontrado");
+        }
     }
 
-    public Optional<Topico> getTopicoById (Long id){
-        return repository.findById(id);
+    public TopicoDTO getTopicoById (Long id) throws EntityNotFoundException {
+        Topico topico = repository.findById(id).orElseThrow();
+        TopicoDTO topicoDTO = modelMapper.map(topico,TopicoDTO.class);
+        return topicoDTO;
     }
 
-    public Set<Respuesta> getRespuestasTopico (long id) {
-        return repository.findById(id).get().getRespuestas();
-    }
+
+
 }
