+++ src/main/java/com/alura/foro/service/UsuarioService.java
@@ -1,12 +1,19 @@
 package com.alura.foro.service;
 
+import com.alura.foro.dto.responce.UsuarioDTO;
+import com.alura.foro.exceptions.BadRequestException;
 import com.alura.foro.modelo.Usuario;
 import com.alura.foro.repository.UsuarioRepository;
-import jakarta.validation.Valid;
+import jakarta.persistence.EntityNotFoundException;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class UsuarioService {
@@ -14,21 +21,48 @@ public class UsuarioService {
     @Autowired
     private UsuarioRepository repository;
 
+    @Autowired
+    private ModelMapper modelMapper;
 
-    public Usuario saveUsuarios(@Valid Usuario usuario){
-        return repository.save(usuario);
+    public UsuarioDTO setUsuario (Usuario usuario)throws BadRequestException {
+        if (usuario.getId()== null){
+            Usuario saveUser = repository.save(usuario);
+            UsuarioDTO usuarioDTO = modelMapper.map(saveUser, UsuarioDTO.class);
+            return usuarioDTO;
+        } else {
+            throw new BadRequestException("Usuario existente");
+        }
+
     }
 
-    public List<Usuario> getUsuarios(){
-        return repository.findAll();
+
+    public List<UsuarioDTO> getUsuarios(){
+
+        List<Usuario> usuarios = repository.findAll();
+
+        return usuarios.stream().map(UsuarioDTO::new).collect(Collectors.toList());
+
+
     }
 
-    public void deleteUsuario (Long id){
-        repository.deleteById(id);
+    public void deleteUsuario (Long id) throws EntityNotFoundException {
+        if (repository.findById(id).isPresent()){
+            repository.deleteById(id);
+        }else{
+            throw new EntityNotFoundException("Usuario no encontrado");
+        }
+
+
+
     }
 
-    public void updateUsuario (Usuario usuario){
-        repository.save(usuario);
-    }
+    public Optional<Usuario> findUserById (Long id) {return repository.findById(id);}
 
+    public void updateUsuario (Usuario usuario)throws EntityNotFoundException {
+        if (repository.findById(usuario.getId()).isPresent()) {
+            repository.save(usuario);
+        } else {
+            throw new EntityNotFoundException("Usuario no encontrado");
+        }
+    }
 }
